Move monitor detection to extra file
In order to hide X11 specific things, move xinerama specific code to
extra file.
